A lot of stuff Reformat code Make bundle.project.json automatically fill itself Update license notice for ChromaStringRenderer Change command to Essential API command update gradle
