+++ gradle.properties
@@ -3,7 +3,6 @@ modVersion=2.2
 modBaseName=HeightLimitMod
 forgeVersion=1.8.9-11.15.1.2318-1.8.9
 mcpVersion=stable_22
-tweakClass=com.pinkulu.forge.FMLLoadingPlugin
 
 org.gradle.daemon=true
 org.gradle.parallel=true
