+++ src/main/java/com/pinkulu/hlm/gui/renderHightLimit/PositionConfig.java
@@ -1,0 +1,13 @@
+package com.pinkulu.hlm.gui.renderHightLimit;
+
+public class PositionConfig {
+
+    public static Double BlocksTillMaxX = 0.5;
+    public static Double BlocksTillMaxY = 0.7;
+
+    public static Double MaxHeightX = 0.5;
+    public static Double MaxHeightY = 0.5;
+
+    public static Double CurrentMapX = 0.5;
+    public static Double CurrentMapY = 0.3;
+}
