+++ src/main/java/com/pinkulu/hlm/util/DelayedTask.java
@@ -1,4 +1,4 @@
-package com.pinkulu.util;
+package com.pinkulu.hlm.util;
 
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
@@ -7,20 +7,20 @@ import net.minecraftforge.fml.common.gameevent.TickEvent.Phase;
 
 public class DelayedTask {
 
-    private int counter;
     private final Runnable run;
+    private int counter;
 
-    public DelayedTask(Runnable run, int ticks){
+    public DelayedTask(Runnable run, int ticks) {
         counter = ticks;
         this.run = run;
         MinecraftForge.EVENT_BUS.register(this);
     }
 
     @SubscribeEvent
-    public void onTick(TickEvent.ClientTickEvent event){
-        if(event.phase != Phase.START)return;
+    public void onTick(TickEvent.ClientTickEvent event) {
+        if (event.phase != Phase.START) return;
 
-        if(counter <= 0){
+        if (counter <= 0) {
             MinecraftForge.EVENT_BUS.unregister(this);
             run.run();
         }
