+++ src/main/java/com/pinkulu/hlm/util/FileUtil.java
@@ -1,15 +1,16 @@
-package com.pinkulu.util;
+package com.pinkulu.hlm.util;
+
+import org.json.JSONObject;
 
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.util.Scanner;
-import org.json.JSONObject;
 
-public class readFile {
+public class FileUtil {
     public static boolean isInvalid;
     public static int limit;
 
-    public static void read(String gametype, String map){
+    public static void read(String gametype, String map) {
         try {
             File myObj = new File("./config/HeightLimitMod/limits.json");
             Scanner myReader = new Scanner(myObj);
@@ -19,14 +20,12 @@ public class readFile {
                 final_string.append(data);
             }
             JSONObject obj = new JSONObject(String.valueOf(final_string));
-            if(obj.has(gametype.toLowerCase())) {
+            if (obj.has(gametype.toLowerCase())) {
                 if (obj.getJSONObject(gametype.toLowerCase()).has(map.toLowerCase())) {
                     limit = obj.getJSONObject(gametype.toLowerCase()).getInt(map.toLowerCase());
                     isInvalid = false;
-                }
-                else isInvalid = true;
-            }
-            else isInvalid = true;
+                } else isInvalid = true;
+            } else isInvalid = true;
             myReader.close();
         } catch (FileNotFoundException e) {
             System.out.println("An error occurred.");
