+++ src/main/java/com/pinkulu/hlm/util/Replace.java
@@ -1,0 +1,7 @@
+package com.pinkulu.hlm.util;
+
+public class Replace {
+    public static String space(String text) {
+        return text.replaceAll(" ", "_");
+    }
+}
