--- src/main/java/com/pinkulu/util/Replace.java
@@ -1,5 +1,0 @@
-package com.pinkulu.util;
-
-public class Replace {
-    public static String space(String text){return text.replaceAll(" ", "_");}
-}
