+++ src/main/resources/bundle.project.json
@@ -1,6 +1,6 @@
 {
   "id": "heightlimitmod",
-  "version": "2.2",
-  "minecraft_version": "1.8.9",
+  "version": "${version}",
+  "minecraft_version": "${mcversion}",
   "platform": "forge"
 }
